% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{clone_environment}
\alias{clone_environment}
\title{Clone objects in a shared environment}
\usage{
clone_environment(name, envir = parent.frame())
}
\arguments{
\item{name}{character. name of the shared environment.}

\item{envir}{an environment to which the objects in the
shared environment are exported.}
}
\description{
Clone objects in a shared environment
}

