% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{clone_object}
\alias{clone_object}
\title{Clone an object from shared memory}
\usage{
clone_object(name)
}
\arguments{
\item{name}{character. The name of the object.}
}
\description{
Clone an object from shared memory
}
\examples{
\dontrun{
clone_object("shared_numbers")
}
}

