% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clone_objects}
\alias{clone_objects}
\title{Clone objects from shared memory to an environment}
\usage{
clone_objects(..., envir = parent.frame())
}
\arguments{
\item{...}{character values indicating names of objects
in shared memory.}

\item{envir}{an environment to which the cloned objects
are exported.}
}
\description{
Clone objects from shared memory to an environment
}
