% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{get_object}
\alias{get_object}
\title{Copy object from shared memory}
\usage{
get_object(name)
}
\arguments{
\item{name}{character. The name of the object.}
}
\description{
Copy object from shared memory
}
\examples{
\dontrun{
get_object("shared_numbers")
}
}

