% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_object}
\alias{remove_object}
\title{Remove object from shared memory}
\usage{
remove_object(name)
}
\arguments{
\item{name}{character. The name of the object.}
}
\description{
Remove object from shared memory
}
\examples{
\dontrun{
remove_object("shared_data")
}
}

