% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{share_environment}
\alias{share_environment}
\title{Share the objects of an environment to shared memory}
\usage{
share_environment(name, envir = parent.frame(), all.names = TRUE,
  overwrite = TRUE)
}
\arguments{
\item{name}{character. the name of the shared memory object.}

\item{envir}{an environment to share. In default, the calling environment
will be shared.}

\item{all.names}{a logical indicating whether to share all values or
(default) those whose names do not begin with a dot.}

\item{overwrite}{a logical to indicate whether to overwrite
existing memory with the same name.}
}
\description{
Share the objects of an environment to shared memory
}

