% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{share_object}
\alias{share_object}
\title{Share an object to shared memory}
\usage{
share_object(x, name, overwrite = TRUE)
}
\arguments{
\item{x}{an object.}

\item{name}{character. The name of the object.}

\item{overwrite}{logical to indicate whether to overwrite
existing memory with the same name.}
}
\description{
Share an object to shared memory
}
\examples{
\dontrun{
share_object(mtcars, "shared_mtcars")
share_object(rnorm(1000), "shared_rnd_numbers")
}
}

