% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{share_objects}
\alias{share_objects}
\title{Share objects to shared memory}
\usage{
share_objects(...)
}
\arguments{
\item{...}{objects to share. Unnamed arguments must be given as
a symbol rather than an expression.}
}
\description{
Share objects to shared memory
}
