% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unshare}
\alias{unshare}
\title{Remove objects from shared memory}
\usage{
unshare(...)
}
\arguments{
\item{...}{The names of the object.}
}
\description{
Remove objects from shared memory
}
\examples{
\dontrun{
unshare("shared_data")
}
}
