% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clone_object}
\alias{clone_object}
\title{Clone an object from shared memory}
\usage{
clone_object(name, default)
}
\arguments{
\item{name}{character. The name of the object.}

\item{default}{If the shared-object does not exist, this object
will be returned. Missing to suppress this behavior, and a stop
will occur instead.}
}
\description{
Clone an object from shared memory
}
\examples{
\dontrun{
clone_object("shared_numbers")
}
}
