% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{sharing}
\alias{sharing}
\title{Do objects exist in shared memory?}
\usage{
sharing(...)
}
\arguments{
\item{...}{The names of the objects.}
}
\description{
Do objects exist in shared memory?
}
\examples{
\dontrun{
sharing("obj1")
}
}

